More error handling, fine tuning large values and non overwriting on initialization
